==> 3_Joining_Datasets_and_data_cleaning.ipynb <==
## Extracting the dataset
Gathering the datasets resulting from the first and second scripts

import os
import fnmatch
import sys
import shutil
import re
import pandas as pd
from sklearn.metrics import *
from scipy import stats
import numpy as np

ai_densities = "df_ai_densities.json"
radiologist_densities = "df_radiologist.json"
density_values = {"A":1, "B":2, "C":3, "D":4}

df_ai_densities = pd.read_json(ai_densities)
df_ai_densities = df_ai_densities.replace(density_values)

df_radiologist_densities = pd.read_json(radiologist_densities)
df_radiologist_densities = df_radiologist_densities.replace(density_values)

full_df= df_radiologist_densities.merge(df_ai_densities,how='inner', left_on='process', right_index=True)
full_df.drop(columns=["process"],inplace= True)

## Final output - Full dataset for comparison and evaluation of metrics

full_df.to_json(r'full_df.json')
