==> 4_AI_and_Radiologist_Comparison.ipynb <==
## AI and Radiologist Comparison
Final script for metric extraction and performance evaluation

import os
import fnmatch
import sys
import shutil
import re
import pandas as pd
from sklearn.metrics import *
from scipy import stats
import numpy as np

df_load = "full_df.json"
full_df = pd.read_json(df_load)
full_df.reset_index(drop=True, inplace=True)

## Dataset

print(f" A total of {full_df.shape[0]-1} unique patients and {full_df.shape[0]*4} (MLO and CC) views were evaluted.")

## Creating dichotomous variable (Non-dense vs Dense)

#create column for non-dense or dense breast 
def RAD_dense_non_dense(x):
    if x['density'] <3 :
        return 0
    else:
        return 1
    
def AI_dense_non_dense(x):
    if x['aidensity'] <3 :
        return 0
    else:
        return 1

full_df['radiologist_2_densities']=full_df.apply(lambda x : RAD_dense_non_dense(x),axis=1)    
full_df['ai_2_densities']=full_df.apply(lambda x : AI_dense_non_dense(x),axis=1)

## Computing metrics by Radiologist

### Non-dense vs dense category

accuracy2=[]
kappa2=[]
for i in range(5):
    new_df = full_df[full_df['radiologist']==i]
    y_true = new_df["radiologist_2_densities"]
    y_pred = new_df["ai_2_densities"]
    accuracy2.append(accuracy_score(y_true, y_pred))
    kappa2.append(cohen_kappa_score(y_true, y_pred))
    print(f"Using 2 density scale: Annotations from Radiologist {i}: Accuracy is :{accuracy_score(y_true, y_pred)} and cohen kappa is:{cohen_kappa_score(y_true, y_pred)}.")

accuracy2CI = stats.t.interval(alpha=0.95, loc=np.mean(accuracy2), df=len(accuracy2)-1, scale=stats.sem(accuracy2)) 
kappa2CI = stats.t.interval(alpha=0.95, loc=np.mean(kappa2), df=len(kappa2)-1, scale=stats.sem(kappa2)) 


print(f"\n2 Density scale mean accuracy:{np.mean(accuracy2)} and 95% C.I. {accuracy2CI}.") 
print(f"2 Density scale mean cohen kappa:{np.mean(kappa2)} and 95% C.I. {kappa2CI }.")  
print(f"2 Density scale - accuracy standard deviation:{np.std(accuracy2)}.") 

### Four BI-RADS density categories

accuracy4=[]
kappa4=[]
for i in range(5):
    new_df = full_df[full_df['radiologist']==i]
    y_true_4scale = new_df["density"].values
    y_pred_4scale = new_df["aidensity"].values
    accuracy4.append(accuracy_score(y_true_4scale, y_pred_4scale))
    kappa4.append(cohen_kappa_score(y_true_4scale, y_pred_4scale))
    print(f"Using 4 density scale: Annotations from Radiologist {i}:Accuracy is :{accuracy_score(y_true_4scale, y_pred_4scale)} and cohen kappa is:{cohen_kappa_score(y_true_4scale, y_pred_4scale)}.")

accuracy4CI = stats.t.interval(alpha=0.95, loc=np.mean(accuracy4), df=len(accuracy4)-1, scale=stats.sem(accuracy4)) 
kappa4CI = stats.t.interval(alpha=0.95, loc=np.mean(kappa4), df=len(kappa4)-1, scale=stats.sem(kappa4)) 
    
print(f"\n4 Density scale mean accuracy:{np.mean(accuracy4)}, and 95% C.I. {accuracy4CI }.") 
print(f"4 Density scale mean cohen kappa:{np.mean(kappa4)} and 95% C.I. {kappa4CI }.")  
print(f"4 Density scale - accuracy standard deviation:{np.std(accuracy4)}.")
